pyproject.toml
[project]
name = "wattameter"
version = "0.4.1"
description = "Default template for PDM package"
authors = [
    {name = "Weslley da Silva Pereira", email = "wdasilv@nrel.gov"},
]
dependencies = [
    "codecarbon>=3.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "BSD-3-Clause"}

[project.scripts]
wattameter = "wattameter.cli.main:main"
wattameter_powerlog_filename = "wattameter.cli.main:print_powerlog_filename"
wattameter_emissions_filename = "wattameter.cli.main:print_emissions_filename"

[dependency-groups]
example = [
    "jupyter",
    "matplotlib",
    "py-cpuinfo",
    "nvidia-ml-py",
]
test = [
    "pytest",
    "pytest-asyncio",
]

[tool.pdm]
distribution = true
package-dir = "src"
includes = ["src/wattameter/utils/*.sh"]

[tool.pdm.scripts]
devwattameter = { call = "wattameter.cli.main:main" }

[tool.pytest.ini_options]
testpaths = ["tests"]
